fix: 再起動ダイアログで断るとエンジンがアップデートされない問題を修正 (#1356)

+++ src/background.ts
@@ -923,23 +923,13 @@ app.on("before-quit", async (event) => {
   const numLivingEngineProcess = Object.entries(killingProcessPromises).length;
 
   // すべてのエンジンプロセスが停止している
-  if (numLivingEngineProcess === 0) {
-    log.info(
-      "All ENGINE processes are killed, running post engine kill process"
-    );
-    if (appState.willRestart) {
-      // awaitする前にevent.preventDefault()を呼び出さないとアプリがそのまま終了してしまう
-      event.preventDefault();
-    }
-
-    // エンジン終了後の処理を実行
-    await vvppManager.handleMarkedEngineDirs();
-
+  if (numLivingEngineProcess === 0 && !vvppManager.hasMarkedEngineDirs()) {
     if (appState.willRestart) {
       // 再起動フラグが立っている場合はフラグを戻して再起動する
       log.info(
         "Post engine kill process done. Now restarting app because of willRestart flag"
       );
+      event.preventDefault();
 
       appState.willRestart = false;
       appState.willQuit = false;
@@ -980,10 +970,14 @@ app.on("before-quit", async (event) => {
   // すべてのエンジンプロセスキル処理が完了するまで待機
   await Promise.all(waitingKilledPromises);
 
-  // アプリケーションの終了を再試行する
+  // エンジン終了後の処理を実行
   log.info(
-    "All ENGINE process kill operations done. Attempting to quit app again"
+    "All ENGINE process kill operations done. Running post engine kill process"
   );
+  await vvppManager.handleMarkedEngineDirs();
+
+  // アプリケーションの終了を再試行する
+  log.info("Attempting to quit app again");
   app.quit();
   return;
 });
